# Performance improvements for GTPv2