* store peer nodes' address as a string to avoid calling String() every time looking up Session by TEID

https://gist.github.com/wmnsk/1de79b7cc1852c3020596f5f0f3a04fb

```
BenchmarkAddrComparisonString1-2         4453245               241 ns/op
BenchmarkAddrComparisonString2-2        360841701                3.17 ns/op
BenchmarkAddrComparisonBytes-2          286502511                4.12 ns/op
BenchmarkAddrComparisonReflect-2          639075              1600 ns/op
```

* ... to be updated?